Finish writing tests
